Allow float height in releasefile

+++ src/common/releasefile.f90
@@ -55,7 +55,7 @@ subroutine  releasefile(filename, release1)
   integer :: ifd, ios, iexit, nlines
   integer :: i
   real :: hour, lasthour
-  integer :: height
+  real :: height
   integer :: ihour, iheight, icmp
   real ::    rel_s
   character(32) :: comp
